update code of benchmark the language model
add support for loading data from different data sources: postgres, csv 

fix the issue of the inference is not actually executed when feed dataset object to the pipeline, instead use the data loader.
